fix 64 bit overflow

+++ src/encoding.c
@@ -35,7 +35,7 @@ uint64_t decode_64bit(uint8_t *byetarray)
     uint64_t decoded = 0;
     uint8_t shift = 56;
     for (uint16_t i = 0; i < 8; i++) {
-        decoded = decoded | (byetarray[i] << shift);
+        decoded = decoded | ((uint64_t)byetarray[i] << shift);
         shift -= 8;
     }
     return decoded;
